@@ -2,9 +2,11 @@ name: objectid_example
 description: ObjectId example.
 version: 1.0.0
 publish_to: none
+executables: 
+  example: ./bin/example.dart
 
 environment:
-  sdk: ">=2.8.1 <3.0.0"
+  sdk: ">=2.12.0 <4.0.0"
 
 dependencies:
   benchmark_harness: any
@@ -13,4 +15,5 @@ dependencies:
     path: ../
 
 dev_dependencies:
-  pedantic: ^1.9.0
+  lints: ^2.1.0
+
